feat: introduce Quick Share functionality

+++ utils/Icons.qml
@@ -166,6 +166,8 @@ Singleton {
 
     function getNotifIcon(summary: string, urgency: int): string {
         summary = summary.toLowerCase();
+        if (summary.includes("quickshare"))
+            return "near_me";
         if (summary.includes("reboot"))
             return "restart_alt";
         if (summary.includes("recording"))
